Update dataflow for multi-node, add humidity, null
Update the /control data flow to accept optional temp, pressure and humidity
readings for nodes. Non-specified or unparseable values are null.

The DB schema has been altered to keep the readings from the nodes in a separate
table to non-environmental historical data, like whether the furnace is on or
people are home.